handler does articles and eventDocuments
