+++ pipeline/src/indices/events.ts
@@ -1,0 +1,1 @@
+export const index = "events";
