+++ pipeline/src/transformers/article.ts
@@ -81,11 +81,6 @@ function transformLabelType(
     : (defaultArticleFormat as ArticleFormat);
 }
 
-export const isArticle = (
-  doc: prismic.PrismicDocument
-): doc is ArticlePrismicDocument =>
-  ["articles", "webcomics"].includes(doc.type);
-
 export const transformArticle = (
   document: ArticlePrismicDocument
 ): ElasticsearchArticle => {
